+++ Optimizer/Options.cs
@@ -29,12 +29,17 @@ namespace Optimizer
         public bool DisableHomeGroup { get; set; }
         public bool DisableSuperfetch { get; set; }
         public bool DisableTelemetryTasks { get; set; }
-        public bool DisableOffice2016Telemetry { get; set; }
         public bool DisableCompatibilityAssistant { get; set; }
         public bool DisableFaxService { get; set; }
         public bool DisableSmartScreen { get; set; }
         public bool DisableCloudClipboard { get; set; }
         public bool DisableStickyKeys { get; set; }
+
+        // apps telemetry
+        public bool DisableOffice2016Telemetry { get; set; }
+        public bool DisableVisualStudioTelemetry { get; set; }
+        public bool DisableFirefoxTemeletry { get; set; }
+        public bool DisableChromeTelemetry { get; set; }
 
         // windows 10
         public bool EnableLegacyVolumeSlider { get; set; }